requirements.txt
streamlit==1.32.2
pandas==2.1.3
numpy==1.24.3
plotly==5.18.0
matplotlib==3.8.2
reportlab==4.0.9
scikit-learn==1.3.2
cryptography==41.0.7
argon2-cffi==23.2.0

